% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adios.r
\docType{data}
\name{adios}
\alias{adios}
\alias{adios_file_R6}
\title{adios_file_R6}
\format{
An object of class \code{adios_R6} (inherits from \code{R6}) of length 3.
}
\usage{
adios
}
\description{
R6 classes and methods for ADIOS2 file operations.
}
\examples{
\dontrun{
suppressMessages(library(hola))

f = adios$file("/path/to/my/dataset")
myvar_step_0 = f$read("my_variable")
f$advance()
myvar_step_2 = f$read("my_variable")
}

}
\keyword{datasets}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{adios_file_R6$new()}}
\item \href{#method-close}{\code{adios_file_R6$close()}}
\item \href{#method-print}{\code{adios_file_R6$print()}}
\item \href{#method-advance}{\code{adios_file_R6$advance()}}
\item \href{#method-read}{\code{adios_file_R6$read()}}
\item \href{#method-clone}{\code{adios_file_R6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_file_R6$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{File path.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Open the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_file_R6$close()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Close the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_file_R6$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print some basic file information.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-advance"></a>}}
\if{latex}{\out{\hypertarget{method-advance}{}}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_file_R6$advance(timeout = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Timeout in seconds between trying to advance the step in
the case data is being streamed.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Advance the step.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_file_R6$read(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{String name of the desired variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Reads the array of a variable into R memory.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_file_R6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
