% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advance.r
\name{adios_advance}
\alias{adios_advance}
\title{advance}
\usage{
adios_advance(af, timeout = 10)
}
\arguments{
\item{af}{An ADIOS2 file; the return of the \code{adios_open()} function.}

\item{timeout}{TODO}
}
\value{
TODO
}
\description{
Advance the current step.
}
