% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_variables.r
\name{adios_available_variables}
\alias{adios_available_variables}
\title{available_variables}
\usage{
adios_available_variables(af)
}
\arguments{
\item{af}{An ADIOS2 file; the return of the \code{adios_open()} function.}
}
\value{
TODO
}
\description{
Show available variables in ADIOS2-managed file.
}
