% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.r
\name{adios_close}
\alias{adios_close}
\title{adios_close}
\usage{
adios_close(af)
}
\arguments{
\item{af}{An ADIOS2 file; the return of the \code{adios_open()} function.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Close an open ADIOS2 file.
}
