% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.r
\name{adios_open}
\alias{adios_open}
\title{adios_open}
\usage{
adios_open(filename, engine)
}
\arguments{
\item{filename}{TODO}

\item{engine}{Either left blank/missing, or one of "bpfile", "dataman", "hdf5",
"insitumpi", or "sst", corresponding to the input. If missing, then we try
to intuit the engine based on the extension in \code{filename}.}
}
\value{
An object of class "adios_file".
}
\description{
Open a file.
}
