% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{adios_read}
\alias{adios_read}
\title{adios_read}
\usage{
adios_read(af, variable)
}
\arguments{
\item{af}{An ADIOS2 file; the return of the \code{adios_open()} function.}

\item{variable}{The variable to read.}
}
\value{
Returns the data: a numeric vector, matrix, or multi-way array.
}
\description{
Read a file opened by \code{adios_open()}.
}
