% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hola-package.r
\docType{package}
\name{hola-package}
\alias{hola-package}
\title{hola}
\description{
The 'ADIOS2' library is a high-performance I/O framework commonly
used in 'HPC' environments by simulation scientists. The 'hola' (helpful
operating link to ADIOS) package is an interface to the this I/O framework.
It supports several file formats, including 'HDF5' and its native BP, as
well as streaming data. It also 'natively' supports n-dimensional arrays and
time series. This enables R to be used 'in-situ' with scientific simulation
data. The package requires an installation of 'ADIOS2' library, whose path
can be specified at configure time, but is otherwise easily installed in the
usual manner.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
