% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adios.r
\name{adios}
\alias{adios}
\alias{adios_R6}
\title{adios_file_R6}
\usage{
adios(path, config, io_name, comm = NULL, mode = "read")
}
\arguments{
\item{path}{File path.}

\item{config}{An ADIOS2 config file.}

\item{io_name}{The IO configuration ("io name" key) to use.}

\item{comm}{A communicator number from pbdMPI if the package is installed
with MPI support; will use \code{MPI_COMM_WORLD} by default. Otherwise
the param is ignored.}

\item{mode}{\code{"read"} or \code{"write"}.}
}
\description{
R6 classes and methods for ADIOS2 file operations.
}
\examples{
\dontrun{
suppressMessages(library(hola))

ad = adios("/path/to/my/dataset")
myvar_step_0 = ad$read("my_variable")
ad$advance()
myvar_step_2 = ad$read("my_variable")
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{adios_R6$new()}}
\item \href{#method-available_variables}{\code{adios_R6$available_variables()}}
\item \href{#method-close}{\code{adios_R6$close()}}
\item \href{#method-fileinfo}{\code{adios_R6$fileinfo()}}
\item \href{#method-print}{\code{adios_R6$print()}}
\item \href{#method-advance}{\code{adios_R6$advance()}}
\item \href{#method-step}{\code{adios_R6$step()}}
\item \href{#method-read}{\code{adios_R6$read()}}
\item \href{#method-read_to_buf}{\code{adios_R6$read_to_buf()}}
\item \href{#method-write}{\code{adios_R6$write()}}
\item \href{#method-clone}{\code{adios_R6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$new(path, config, io_name, comm = NULL, mode = "read")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{File path.}

\item{\code{config}}{An ADIOS2 config file.}

\item{\code{io_name}}{The IO configuration ("io name" key) to use.}

\item{\code{comm}}{A communicator number from pbdMPI if the package is installed
with MPI support; will use \code{MPI_COMM_WORLD} by default. Otherwise
the param is ignored.}

\item{\code{mode}}{\code{"read"} or \code{"write"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Open the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-available_variables"></a>}}
\if{latex}{\out{\hypertarget{method-available_variables}{}}}
\subsection{Method \code{available_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$available_variables()}\if{html}{\out{</div>}}
}

\subsection{Details}{
List available variables in a file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$close()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Close the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fileinfo"></a>}}
\if{latex}{\out{\hypertarget{method-fileinfo}{}}}
\subsection{Method \code{fileinfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$fileinfo()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns basic file info.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print some basic file information.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-advance"></a>}}
\if{latex}{\out{\hypertarget{method-advance}{}}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$advance(timeout = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Timeout in seconds between trying to advance the step in
the case data is being read in streams (otherwise ignored).}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Advance the step when using time steps for reading/writing.
The function returns \code{FALSE} if the step status is not "OK" (e.g.,
when reading but there is nothing left to read), and \code{TRUE}
otherwise.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$step()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the current time step.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$read(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{String name of the desired variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Reads the array of a variable into R memory.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_to_buf"></a>}}
\if{latex}{\out{\hypertarget{method-read_to_buf}{}}}
\subsection{Method \code{read_to_buf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$read_to_buf(var, buf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{String name of the desired variable.}

\item{\code{buf}}{Buffer space.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Reads the array of a variable into pre-allocated R memory. If
the buffer is inappropriately sized, terrible things may happen.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$write(var, x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{String name of the desired variable.}

\item{\code{x}}{R vector/matrix object.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Write the R vector/matrix to a file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
