% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adios.r
\name{adios}
\alias{adios}
\alias{adios_R6}
\title{adios_file_R6}
\usage{
adios(config = NULL)
}
\arguments{
\item{config}{An optional ADIOS2 config file.}
}
\description{
R6 classes and methods for ADIOS2 file operations.
}
\examples{
\dontrun{
suppressMessages(library(hola))

ad = adios()
ad$open("/path/to/my/dataset")
myvar_step_0 = ad$read("my_variable")
f$advance()
myvar_step_2 = ad$read("my_variable")
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{adios_R6$new()}}
\item \href{#method-open}{\code{adios_R6$open()}}
\item \href{#method-available_variables}{\code{adios_R6$available_variables()}}
\item \href{#method-close}{\code{adios_R6$close()}}
\item \href{#method-print}{\code{adios_R6$print()}}
\item \href{#method-advance}{\code{adios_R6$advance()}}
\item \href{#method-read}{\code{adios_R6$read()}}
\item \href{#method-clone}{\code{adios_R6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$new(config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{An optional ADIOS2 config file.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Open the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$open(path, engine = NULL, io_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{File path.}

\item{\code{engine}}{Either left blank/missing, or one of "bpfile", "dataman",
"hdf5", "insitumpi", or "sst", corresponding to the input. If missing,
then we try to intuit the engine based on the extension in
\code{path}.}

\item{\code{io_name}}{The IO configuration ("io name" key) to use if using a
config file.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Open the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-available_variables"></a>}}
\if{latex}{\out{\hypertarget{method-available_variables}{}}}
\subsection{Method \code{available_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$available_variables()}\if{html}{\out{</div>}}
}

\subsection{Details}{
List available variables in a file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$close()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Close the file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print some basic file information.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-advance"></a>}}
\if{latex}{\out{\hypertarget{method-advance}{}}}
\subsection{Method \code{advance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$advance(timeout = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timeout}}{Timeout in seconds between trying to advance the step in
the case data is being streamed.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Advance the step.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$read(var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{String name of the desired variable.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Reads the array of a variable into R memory.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{adios_R6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
